Move init_genesis_store to era/store_init.rs
Separate store initialization from the ERA consumer since it's not part
of the production beacon node path.
